+++ CHANGELOG.md
@@ -1,9 +1,30 @@
+# v1.0.0-rc.4
+## 10/29/2015
+
+1. [](#bugfix)
+    * Fixed a fatal error if you have a collection with missing or invalid `@page: /route`
+
+# v1.0.0-rc.3
+## 10/29/2015
+
+1. [](#new)
+    * New Page collection options! `@self.parent, @self.siblings, @self.descendants` + more
+    * Whitelist of file types for fallback route functionality (images by default)
+1. [](#improved)
+    * Assets switched from defines to streams
+1. [](#bugfix)
+    * README.md typos fixed
+    * Fixed issue with routes that have lang string in them (`/en/english`)
+    * Trim strings before validation so whitespace is not satisfy 'required'
+
 # v1.0.0-rc.2
-## xx/xx/2015
+## 10/27/2015
 
 1. [](#new)
     * Added support for CSS Asset groups
     * Added a `wrapped_site` system option for themes/plugins to use
+    * Pass `Page` object as event to `onTwigPageVariables()` event hook
+    * New `Data.items()` method to get all items
 1. [](#improved)
     * Missing pipelined remote asset will now fail quietly
     * More reliably handle inline JS and CSS to remove only surrounding HTML tags
@@ -11,8 +32,10 @@
     * Improved Medium metadata merging to allow for automatic title/alt/class attributes
     * Moved Grav object to global variable rather than template variable (useful for macros)
     * German language improvements
+    * Updated bundled composer
 1. [](#bugfix)
     * Accept variety of `true` values in `User.authorize()` method 
+    * Fix for `Validation` throwing an error if no label set
 
 # v1.0.0-rc.1
 ## 10/23/2015